+++ KeyboardKitDemoKeyboard/KeyboardViewController.swift
@@ -102,12 +102,6 @@ class KeyboardViewController: KeyboardInputViewController {
     lazy var autocompleteProvider = DemoAutocompleteSuggestionProvider()
     
     lazy var autocompleteToolbar: AutocompleteToolbar = {
-        let proxy = textDocumentProxy
-        let toolbar = AutocompleteToolbar(
-            //buttonCreator: { DemoAutocompleteLabel(word: $0, proxy: proxy) }
-            buttonCreator: { AutocompleteToolbarLabel(text: $0, textDocumentProxy: proxy) }
-        )
-        toolbar.update(with: ["foo", "bar", "baz"])
-        return toolbar
+        AutocompleteToolbar(textDocumentProxy: textDocumentProxy)
     }()
 }