+++ RELEASE_NOTES.md
@@ -8,7 +8,7 @@ This version adds more autocomplete functionality:
 * `AutocompleteToolbar` has a new convenience initializer that makes it even easier to setup autocomplete.
 * `AutocompleteToolbarLabel` is the default autocomplete item view and can be tapped to send text to the text document proxy.
 * `AutocompleteToolbarLabel` behaves like the native iOS autocomplete view and displays centered text until the text must scroll.
-* Autocomplete no longer requires the bugfix timer to work. Instead, just let the action handler request autocomplete suggestions.
+* Autocomplete no longer requires the bugfix timer. Instead, just let the action handler request autocomplete suggestions when a tap action is triggered.
 
 The `StandardKeyboardActionHandler` has new functionality:
 * `animationButtonTap()` - can be overridden to change the default animation of tapped buttons.